BACKPORT: tcp: record most recent RTT in RACK loss detection
Record the most recent RTT in RACK. It is often identical to the
"ca_rtt_us" values in tcp_clean_rtx_queue. But when the packet has
been retransmitted, RACK choses to believe the ACK is for the
(latest) retransmitted packet if the RTT is over minimum RTT.

This requires passing the arrival time of the most recent ACK to
RACK routines. The timestamp is now recorded in the "ack_time"
in tcp_sacktag_state during the ACK processing.

This patch does not change the RACK algorithm itself. It only adds
the RTT variable to prepare the next main patch.

saalim :- adapt tcp_rate_gen to recent changes done
* original https://github.com/danascape/kernel-msm-4.14/commit/deed7be78f512d003c6290da0a781479b31b3d74#diff-0d3389ce2dcd134567fb4d6fa05c89016f7922a044391ef933f79cd265edc4d3R3719

Signed-off-by: Yuchung Cheng <ycheng@google.com>
Signed-off-by: Neal Cardwell <ncardwell@google.com>
Acked-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: danascape <saalimquadri1@gmail.com>
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>
Signed-off-by: Sorayukii <sorayukii69@gmail.com>
Signed-off-by: Yuppi <78759550+Zymboe11@users.noreply.github.com>