net-tcp_bbr: v2: rename bw_hi to max_bw
This is so we can, in the next patch, separate the notion of the
estimated max safe pacing rate (bw_hi) from the notion of the raw
windowed max bw sampled recently (max_bw[0], max_bw[1]).

The motivation is to allow sending rate to match available bw in
datacenter incast with ECN as a congestion signal. This separation is
important because probing for bandwidth when the bottleneck is full
causes the measured bw to be higher than the long-term sustained
available bw. In order to match the aggregate sending rate to the
available bandwidth, the senders will need to adapt their max sending
rate (bw_hi), as well as typical sending rate (bw_lo).

Change-Id: I05a2f3886079360549c22e14914cf21fae72db1b
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>
Signed-off-by: Sorayukii <sorayukii69@gmail.com>
Signed-off-by: Yuppi <78759550+Zymboe11@users.noreply.github.com>